fixes of operations and slice for update schedule
